Core/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
Jinja2==3.1.5
msgpack==1.1.0
redis==5.2.1
requests==2.32.3
Twisted==24.11.0
txaio==23.1.1
whitenoise==6.9.0
PyJWT==2.10.1
Pillow==10.2.0  # (or the latest version)
django-channels-jwt-auth-middleware
dj-database-url
psycopg2-binary
python-dotenv